simplifying api interface, refactoring, and making endpoint more testable
