+++ libs/common/src/common/types/errors.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class ValidationError(BaseModel):
+    message: str
